Fix Zoo 2.2 assignment corrections per checklist
- ICageable: change XPosition/YPosition from double to int
- Animal: reorder constructor (age first), add no-age constructor, change movement props to int, age-based DisplaySize/ResourceKey, fix random init ranges
- Mammal/Bird/Fish: add no-age constructors, fix Move() boundary check pattern (check before add, not add then check)
- All animal subclasses: reorder constructors, add no-age constructors, update age-based DisplaySize overrides (Shark/Platypus/Hummingbird/Ostrich)
- Chimpanzee/Kangaroo/Squirrel: add base.Move() call for position tracking
- AnimalFactory: update all constructor calls to new parameter order
- Guest: change XPosition/YPosition to int, initialize in constructor, update ToString() to include adopted animal name
- MoneyCollectingBooth/GivingBooth: throw NullReferenceException wrapping inner exception (not MissingItemException)
- Cage: use System.Type instead of AnimalType enum for AnimalType property
- Zoo: create cages using typeof() instead of AnimalType enum loop, FindCage compares Type directly
- CageWindow.xaml: replace Canvas with Grid (cageGrid) inside Border
- CageWindow.xaml.cs: new GetViewBox with proportional scaling formula using cageGrid dimensions
- App.xaml: add Baby and Adult suffixed resources for all 9 animal types
- ConsoleHelper: show "Zoo is out of tickets" in NullReferenceException catch block

+++ CagedItems/ICageable.cs
@@ -8,12 +8,12 @@ namespace CagedItems
         /// <summary>
         /// Gets or sets the horizontal position of the item.
         /// </summary>
-        double XPosition { get; set; }
+        int XPosition { get; set; }
 
         /// <summary>
         /// Gets or sets the vertical position of the item.
         /// </summary>
-        double YPosition { get; set; }
+        int YPosition { get; set; }
 
         /// <summary>
         /// Gets the display size of the item.
